move Shell out of repo, remove unused components
