--- components/shell/pages/EmptyPage.tsx
@@ -1,36 +1,0 @@
-'use client'
-
-import * as Icons from '@phosphor-icons/react/dist/ssr'
-import { Link } from '@radix-ui/themes'
-import { useRouter } from 'next/navigation'
-
-import { PageWrapper } from '@/components/pages/PageWrapper'
-import { appConfig } from '@/config/config'
-
-export const EmptyPage = () => {
-  const router = useRouter()
-  return (
-    <PageWrapper className="flex flex-col">
-      <div className="m-auto flex flex-col items-center px-5">
-        <Icons.Cat weight="thin" className="size-60 shrink-0 text-accentA-11" />
-
-        <div
-          className="p-3 text-center text-2xl"
-          style={{ fontFamily: 'var(--font-chakra-petch)' }}
-        >
-          There doesn&apos;t appear to be anything at this address.
-        </div>
-
-        <div className="flex gap-4 p-2">
-          <Link underline="always" color="brown" href="#" onClick={() => router.back()}>
-            Go Back
-          </Link>
-
-          <Link underline="always" color="brown" href={appConfig.baseUrl}>
-            Go Home
-          </Link>
-        </div>
-      </div>
-    </PageWrapper>
-  )
-}
