--- components/shell/pages/index.ts
@@ -1,6 +1,0 @@
-export * from './EditThreadTitle'
-export * from './ModelPicker'
-export * from './ThreadConfig'
-export * from './UserThreads'
-export * from './DeleteThread'
-export * from './CreateThread'
