+++ cosmos.imports.ts
@@ -5,8 +5,7 @@ import { RendererConfig, UserModuleWrappers } from 'react-cosmos-core';
 
 import * as fixture0 from './components/text-document-editor/__fixtures__/TextDocumentEditor';
 import * as fixture1 from './components/composer/__fixtures__/Composer';
-import * as fixture2 from './components/ui/TextareaAutosize.fixture';
-import * as fixture3 from './components/navigation/Navigation.fixture';
+import * as fixture2 from './components/navigation/Navigation.fixture';
 
 import * as decorator0 from './components/cosmos.decorator';
 import * as decorator1 from './components/text-document-editor/__fixtures__/cosmos.decorator';
@@ -19,8 +18,7 @@ export const rendererConfig: RendererConfig = {
 const fixtures = {
   'components/text-document-editor/__fixtures__/TextDocumentEditor.tsx': { module: fixture0 },
   'components/composer/__fixtures__/Composer.tsx': { module: fixture1 },
-  'components/ui/TextareaAutosize.fixture.tsx': { module: fixture2 },
-  'components/navigation/Navigation.fixture.tsx': { module: fixture3 }
+  'components/navigation/Navigation.fixture.tsx': { module: fixture2 }
 };
 
 const decorators = {
